+++ src/djlint/reformat.py
@@ -8,7 +8,10 @@ from __future__ import annotations
 import difflib
 from typing import TYPE_CHECKING
 
-from djlint.formatter.class_attributes import restore_class_attribute_newlines
+from djlint.formatter.class_attributes import (
+    restore_class_attribute_newlines,
+    restore_verbatim_attribute_newlines,
+)
 from djlint.formatter.compress import compress_html
 from djlint.formatter.condense import clean_whitespace, condense_html
 from djlint.formatter.expand import expand_html
@@ -54,6 +57,8 @@ def formatter(config: Config, rawcode: str) -> str:
 
     if config.preserve_class_newlines:
         beautified_code = restore_class_attribute_newlines(beautified_code)
+
+    beautified_code = restore_verbatim_attribute_newlines(beautified_code)
 
     beautified_code = restore_unformatted_blocks(
         beautified_code, unformatted_blocks